Grid suply price changes according to the hour.

+++ src/Web/EnergySystem.Web.ViewModels/Property/PropertyInListViewModel.cs
@@ -12,6 +12,6 @@
 
         public string Address { get; set; }
 
-        public float ElectricityNeed { get; set; }
+        public decimal ElectricityNeed { get; set; }
     }
 }
